feat(admin): user + role creation as modern dialogs
Mirror the new-tenant dialog: user and role creation now open well-designed
popup dialogs from their list "New …" buttons instead of separate pages.
Same fields/validation/mutations (registerUser, upsertRole); on success close,
refresh the list query, and route to the created entity. /users/new and
/roles/new redirect to their lists.

Part of the admin → dashboard design unification (PR #1268).
Build: admin `npm run build` ✓, `eslint` ✓ (0 errors).
